Move cli_executor to ansible.cli
